fixed component file structure
