+++ client/components/ads/AdForm.js
@@ -4,20 +4,13 @@ import PropTypes from 'prop-types'
 import history from '../../history'
 import {
   withStyles,
-  FormLabel,
   Typography,
   Button,
   FormControl,
-  FormControlLabel,
   FormGroup,
   Input,
   InputLabel,
   Card,
-  Radio,
-  RadioGroup,
-  MenuItem,
-  FormHelperText,
-  Select,
   Divider,
   Grid
 } from '@material-ui/core'
