+++ client/components/ads/AdsGalleryGridList.js
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types'
 import { withStyles } from '@material-ui/core/styles'
 import {
   GridListTile,
-  GridListTileBar,
-  IconButton,
   GridList,
   Divider,
   Grid,
